refactor(crypto): Move (Into)CryptoStore traits into a new module
